# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atlas Backend API",
        "description": "API documentation for Atlas Backend - Project Management System",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pg/academic-periods/{academic_period}/phases/{phase}": {
            "get": {
                "tags": [
                    "Phases"
                ],
                "summary": "Get a specific phase",
                "operationId": "205b0e664a05c11e0215bc8307b1c42b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Phase ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phase details with period relationship",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhaseResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Phase not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Phases"
                ],
                "summary": "Update a phase",
                "operationId": "0aaa6f3a8b23bc82c693c08ec2efad06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Phase ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Phase update payload. When updating dates, they must be within the academic period and cannot overlap with other phases.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proyecto de grado I"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phase updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhaseResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Phase not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/pg/academic-periods/{academic_period}/phases/dropdown": {
            "get": {
                "tags": [
                    "Phases"
                ],
                "summary": "Get phases for dropdown (scoped to academic period)",
                "operationId": "bd31034082ea0680657b32f19d7f4e17",
                "parameters": [
                    {
                        "name": "academic_period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pairs ready for selects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "value": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "Proyecto de grado I"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PhaseResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Proyecto de grado I"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-15"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-30"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-15T18:30:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}